Allocate ports in a row instead of randomly
Before ports were randomly picked. Now they are allocated in a row in the same order as in the fleet/gameserver config. This is needed for games like Unturned where the first port is for server list queries, and +1 (the next) port is for in-game traffic. This is hardcoded on the client side and this commit is the only workaround.

Note that tests were not yet modified for this change and will fail.